Move TrustV0 message handlers to authorization mod
