+++ libsplinter/src/network/auth/authorization/trust_v0/handlers.rs
