+++ libsplinter/src/network/auth/handlers/builder.rs
@@ -27,14 +27,14 @@ use crate::network::auth::authorization::challenge::handlers::{
 use crate::network::auth::authorization::trust::handlers::{
     AuthTrustRequestHandler, AuthTrustResponseHandler,
 };
+use crate::network::auth::authorization::trust_v0::handlers::{
+    AuthorizedHandler, ConnectRequestHandler, ConnectResponseHandler, TrustRequestHandler,
+};
 use crate::network::auth::AuthorizationManagerStateMachine;
 use crate::network::auth::ConnectionAuthorizationType;
 use crate::network::dispatch::{ConnectionId, Dispatcher, MessageSender};
 use crate::protos::network::NetworkMessageType;
 
-use super::v0_handlers::{
-    AuthorizedHandler, ConnectRequestHandler, ConnectResponseHandler, TrustRequestHandler,
-};
 #[cfg(any(feature = "trust-authorization", feature = "challenge-authorization"))]
 use super::v1_handlers::{
     builders::{AuthProtocolRequestHandlerBuilder, AuthProtocolResponseHandlerBuilder},
